+++ Week2/2048-game/src/ScoreField.js
@@ -1,5 +1,5 @@
 import React from "react";
-import './App.css';
+import './styles/App.css';
 
 /**
  * Component to display score updates.
